Warning fixes

+++ Code/Engine/Engine/RenderSystem.cpp
@@ -62,7 +62,9 @@ void render::RenderSystem::Update(const core::GameTime& /*gameTime*/)
 				const render::TextureAsset& textureAsset = spriteAsset.m_Texture.get();
 
 				const sf::Texture& texture = textureAsset.m_Texture;
-				const sf::Vector2f size = sf::Vector2f(spriteAsset.m_RectangleSize.x, spriteAsset.m_RectangleSize.y);
+				const sf::Vector2f size = sf::Vector2f(
+					static_cast<float>(spriteAsset.m_RectangleSize.x), 
+					static_cast<float>(spriteAsset.m_RectangleSize.y));
 
 				// #note: inverse scale because of SFML
 				const float scaleX = spriteComponent.m_Size.x / size.x;
@@ -115,7 +117,9 @@ void render::RenderSystem::Update(const core::GameTime& /*gameTime*/)
 				const render::TextureAsset& textureAsset = spriteAsset.m_Texture.get();
 
 				const sf::Texture& texture = textureAsset.m_Texture;
-				const sf::Vector2f size = sf::Vector2f(spriteAsset.m_RectangleSize.x, spriteAsset.m_RectangleSize.y);
+				const sf::Vector2f size = sf::Vector2f(
+					static_cast<float>(spriteAsset.m_RectangleSize.x), 
+					static_cast<float>(spriteAsset.m_RectangleSize.y));
 
 				const float scaleX = flipbookComponent.m_Size.x / size.x;
 				const float scaleY = flipbookComponent.m_Size.y / size.y * -1.f;
